move filenames to be consistent
