+++ pkg/network/ebpf/c/runtime/http.c
@@ -4,7 +4,7 @@
 #include "ipv6.h"
 #include "http.h"
 #include "sockfd.h"
-#include "conn_tuple.h"
+#include "conn-tuple.h"
 
 // TODO: Replace those by injected constants based on system configuration
 // once we have port range detection merged into the codebase.
